requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.3.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
postgrest==0.14.1